requirements.txt
django
djangorestframework
djangorestframework-serializer-extensions
